+++ AliceDiary/src/main/java/com/alice/project/repository/MemberRepository.java
@@ -4,14 +4,12 @@ import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.querydsl.QuerydslPredicateExecutor;
 import org.springframework.stereotype.Repository;
 
 import com.alice.project.domain.Member;
 
 @Repository
-public interface MemberRepository
-		extends JpaRepository<Member, Long>, QuerydslPredicateExecutor<Member>, MemberRepositoryCustom {
+public interface MemberRepository extends JpaRepository<Member, Long> {
 
 	@Query(value = "select member_num from Member where id = :id", nativeQuery = true)
 	Long findMemberNumById(String id);
@@ -36,9 +34,6 @@ public interface MemberRepository
 
 	Member findByName(String name);
 
-	/*
-	 * @Query(value = "select * from MEMBER where ID like '%'||:writer||'%'",
-	 * nativeQuery = true) List<Member> findMemberNumByIdContaining(String writer);
-	 */
-
+	@Query(value = "select * from Member where ID like '%'||:id||'%'", nativeQuery = true)
+	Member findMemberById(String id);
 }
