--- AliceDiary/src/main/java/com/alice/project/repository/ReplyRepositoryCustom.java
@@ -1,15 +1,0 @@
-/*
- * package com.alice.project.repository;
- * 
- * import java.util.List;
- * 
- * import org.springframework.stereotype.Repository;
- * 
- * import com.alice.project.domain.Reply;
- * 
- * @Repository public interface ReplyRepositoryCustom {
- * 
- * List<Reply> findGroupByPostNum(Long postNum);
- * 
- * }
- */
